services/project_service/requirements.txt
flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
PyJWT
python-dotenv
requests
